requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
langchain-ollama
chromadb
pydantic
python-multipart
langgraph-checkpoint-sqlite
ddgs
requests    
beautifulsoup4 
readability-lxml
debugpy